refactor: renaming, docs, and exports more clean
