+++ src/index.ts
@@ -11,9 +11,8 @@ export * as Sern from './handler/sern';
 export * from './types/handler';
 export * from './types/module';
 export * from './types/plugin';
-export * from './handler/structures/structxports';
+export * from './handler/structures';
 export * from './handler/plugins';
 export * from './handler/contracts';
 export { SernEmitter };
-export { useContainerRaw } from './handler/dependencies/provider';
-export { single, transient, many } from './handler/dependencies/lifetimeFunctions';
+export * from './handler/dependencies';
